requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
